Refactor runtime store implementation to streamline codebase
- Removed extensive unused code and definitions from `runtime_store.rs`, significantly reducing file size and improving readability.
- Simplified the structure of the `RuntimeStore` trait and its implementations, focusing on essential functionality.
- Enhanced clarity by consolidating related components and removing legacy code, paving the way for future improvements.
