+++ bitloops/src/host/runtime_store/util.rs
@@ -1,0 +1,6 @@
+use sha2::{Digest, Sha256};
+
+pub(crate) fn sha256_hex(data: &[u8]) -> String {
+    let digest = Sha256::digest(data);
+    digest.iter().map(|byte| format!("{byte:02x}")).collect()
+}
